Fix default actions not being applied on first load
Not my favourite way to solve things, but with our current setup, this
is the way to go.

Ensures we have fetched the acl defaults from the backend
before opening the create event/series modal, so that Formik can be
properly initialized with said defaults
without the need for enabling reinitialization.